requirements.txt
Django~=5.1.1
Pillow~=10.4.0
django-colorfield~=0.11.0